# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /employees/{id}:
    get:
      tags:
      - employee-controller
      operationId: getEmployeeById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EmployeeModel'
    put:
      tags:
      - employee-controller
      operationId: updateEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeModel'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EmployeeModel'
    delete:
      tags:
      - employee-controller
      operationId: deleteEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /employees:
    get:
      tags:
      - employee-controller
      operationId: getAllEmployees
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeModel'
    post:
      tags:
      - employee-controller
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeModel'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EmployeeModel'
components:
  schemas:
    EmployeeModel:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        emailAddress:
          type: string
        phoneNumber:
          type: string
        birthDate:
          type: string
          format: date
        jobTitle:
          type: string
        department:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date
        managerReporting:
          type: string